src/requirements.txt
SQLAlchemy==1.3.22
aiohttp==3.7.3
PyJWT==2.0.1
psycopg2-binary==2.8.6